--- payment-process-service/src/main/java/com/example/auction/models/repositories/BetRepository.java
@@ -1,11 +1,0 @@
-package com.example.auction.models.repositories;
-
-import com.example.auction.models.entities.Bet;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface BetRepository extends JpaRepository<Bet, Long> {
-}
