--- payment-process-service/src/main/java/com/example/auction/models/repositories/UserRepository.java
@@ -1,9 +1,0 @@
-package com.example.auction.models.repositories;
-
-import com.example.auction.models.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-}
