--- payment-process-service/src/main/resources/application.properties
@@ -1,30 +1,0 @@
-logging.level.root=INFO
-spring.application.name=auctionSite
-
-spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
-spring.datasource.username=postgres
-spring.datasource.password=reallyNot10Generated_Passw0rd
-
-spring.jpa.properties.hibernate.jdbc.time_zone=UTC
-spring.jpa.show-sql=false
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.hibernate.ddl-auto=validate
-spring.jpa.open-in-view=false
-javax.persistence.query.timeout=3000
-
-spring.kafka.bootstrap-servers=localhost:29093
-spring.kafka.consumer.group-id=auction-main-service
-spring.kafka.consumer.enable-auto-commit=false
-spring.kafka.listener.ack-mode=manual_immediate
-
-topic-user-creation-request=user-create
-topic-lot-purchase-request=lot-purchase
-topic-lot-purchase-result=lot-purchase-response
-
-finished-lots-check-delay=60000
-outbox-attempt-frequency=PT1S
-outbox-flush-frequency=1000
-
-resilience4j.retry.instances.user-generation.retry-exceptions[0]=Exception
-resilience4j.retry.instances.user-generation.wait-duration=0ms
-resilience4j.retry.instances.user-generation.max-attempts=3
