+++ purchase-service/src/main/java/com/example/auction/PurchaseServiceApplication.java
@@ -1,0 +1,15 @@
+package com.example.auction;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@SpringBootApplication
+@EnableScheduling
+public class PurchaseServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PurchaseServiceApplication.class, args);
+	}
+
+}
