+++ purchase-service/src/main/java/com/example/auction/models/entities/User.java
@@ -1,0 +1,76 @@
+package com.example.auction.models.entities;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import org.hibernate.annotations.Fetch;
+import org.hibernate.proxy.HibernateProxy;
+
+import java.util.*;
+
+import static jakarta.persistence.CascadeType.PERSIST;
+import static org.hibernate.annotations.FetchMode.SUBSELECT;
+
+@Entity
+@Table(name = "account")
+public class User {
+  @Column(updatable = false)
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(name = "main_db_id")
+  private UUID mainServiceId;
+
+  @Embedded
+  private Money money;
+
+  protected User() {
+  }
+
+  public User(UUID mainServiceId) {
+    this.mainServiceId = mainServiceId;
+    this.money = new Money();
+  }
+
+  public Long getId() {
+    return id;
+  }
+
+  protected void setId(Long id) {
+    this.id = id;
+  }
+
+  public UUID getMainServiceId() {
+    return mainServiceId;
+  }
+
+  protected void setMainServiceId(UUID mainServiceId) {
+    this.mainServiceId = mainServiceId;
+  }
+
+  public Money getMoney() {
+    return money;
+  }
+
+  public void setMoney(Money money) {
+    this.money = money;
+  }
+
+  @Override
+  public final boolean equals(Object o) {
+    if (this == o) return true;
+    if (o == null) return false;
+    Class<?> oEffectiveClass = o instanceof HibernateProxy ? ((HibernateProxy) o).getHibernateLazyInitializer().getPersistentClass() : o.getClass();
+    Class<?> thisEffectiveClass = this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass() : this.getClass();
+    if (thisEffectiveClass != oEffectiveClass) return false;
+    User user = (User) o;
+    return getId() != null && Objects.equals(getId(), user.getId());
+  }
+
+  @Override
+  public final int hashCode() {
+    return this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass().hashCode() : getClass().hashCode();
+  }
+}
