+++ purchase-service/src/main/java/com/example/auction/models/enums/LotState.java
