+++ purchase-service/src/main/java/com/example/auction/models/exceptions/LotCreateException.java
