+++ purchase-service/src/main/java/com/example/auction/models/exceptions/RequestAlreadyProcessedException.java
@@ -1,0 +1,7 @@
+package com.example.auction.models.exceptions;
+
+public class RequestAlreadyProcessedException extends RuntimeException {
+  public RequestAlreadyProcessedException(String message) {
+    super(message);
+  }
+}
