+++ purchase-service/src/main/java/com/example/auction/models/repositories/UserRepository.java
@@ -1,0 +1,14 @@
+package com.example.auction.models.repositories;
+
+import com.example.auction.models.entities.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+  Optional<User> findByMainServiceId(UUID mainServiceId);
+}
