+++ purchase-service/src/main/resources/application.properties
@@ -1,0 +1,24 @@
+logging.level.root=INFO
+spring.application.name=purchaseService
+
+spring.datasource.url=jdbc:postgresql://localhost:5432/bank_accounts
+spring.datasource.username=postgres
+spring.datasource.password=reallyNot10Generated_Passw0rd
+
+spring.jpa.show-sql=false
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.open-in-view=false
+javax.persistence.query.timeout=3000
+
+spring.kafka.bootstrap-servers=localhost:29093
+spring.kafka.consumer.group-id=purchase-service
+spring.kafka.consumer.enable-auto-commit=false
+spring.kafka.listener.ack-mode=manual_immediate
+
+topic-user-creation-request=user-create
+topic-lot-purchase-request=lot-purchase
+topic-lot-purchase-result=lot-purchase-response
+
+outbox-attempt-frequency=PT1S
+outbox-flush-frequency=1000
